move logic back to manager.py
